# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Service Portal API",
    "version": "1.0.0",
...
  },
  "paths": {
...
    "/api/v1/service-requests/{request_id}/comments": {
      "get": {
        "tags": [
          "comments"
        ],
        "summary": "List a request's comments",
        "description": "A page of this request's comments the caller may see (CM-1 to CM-4).\n\n``user`` is declared alongside ``parent`` even though the parent dependency\nresolves it too — FastAPI caches a dependency per request, so this is one\nlookup, not two, and the alternative is reading the caller off the parent\nobject, which it is not.",
        "operationId": "list_comments_api_v1_service_requests__request_id__comments_get",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_CommentOut_"
                }
              }
            }
          },
          "422": {
            "description": "Returned when one or more supplied fields are invalid (XC-4).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Returned when the `access_token` cookie is absent, expired, or no longer names an active user — no distinction is drawn between those cases (XC-5).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Returned when no service request with that id is visible to the caller — a malformed id, an id matching no row, and an id belonging to someone else are answered identically by design (XC-7, SR-13).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Returned when the request failed for a reason the API does not attribute to the caller — `message` never carries internal detail and the traceback is logged server-side (XC-14).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessTokenCookie": []
          }
        ]
      },
      "post": {
        "tags": [
          "comments"
        ],
        "summary": "Comment on a request",
        "description": "Post a comment on a request the caller may see (CM-5 to CM-9).\n\n``author_id`` is the authenticated user and is never read from the body\n(XC-8) — the schema has no field for it, so a body supplying one creates\nexactly the same row as one that doesn't.",
        "operationId": "create_comment_api_v1_service_requests__request_id__comments_post",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          },
          {
            "name": "X-Requested-With",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Requested-With",
              "default": "XMLHttpRequest"
            },
            "description": "CSRF guard (XC-9). Any non-empty value. Enforced by middleware before routing, so a request without it is answered `403` without its body or its cookies being read — which is why it is absent from this operation's handler signature."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentOut"
                }
              }
            }
          },
          "403": {
            "description": "Returned when the `X-Requested-With` header is missing, which middleware refuses before routing (XC-9); or when a caller who is not an administrator asks for `is_internal: true` — the comment is refused rather than quietly made public (CM-7).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Returned when one or more supplied fields are invalid (XC-4).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Returned when the `access_token` cookie is absent, expired, or no longer names an active user — no distinction is drawn between those cases (XC-5).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Returned when no service request with that id is visible to the caller — a malformed id, an id matching no row, and an id belonging to someone else are answered identically by design (XC-7, SR-13).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Returned when the request failed for a reason the API does not attribute to the caller — `message` never carries internal detail and the traceback is logged server-side (XC-14).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessTokenCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "CommentCreate": {
        "properties": {
          "body": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 1,
            "title": "Body"
          },
          "is_internal": {
            "type": "boolean",
            "title": "Is Internal",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "body"
        ],
        "title": "CommentCreate",
        "description": "``POST /service-requests/{id}/comments`` body (CM-5 through CM-8)."
      },
      "CommentOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "author": {
            "$ref": "#/components/schemas/UserSummary"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "is_internal": {
            "type": "boolean",
            "title": "Is Internal"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "author",
          "body",
          "is_internal",
          "created_at",
          "updated_at"
        ],
        "title": "CommentOut",
        "description": "One comment, as design.md §6 writes it.\n\n``is_internal`` is on every comment, including the ones a `user`-role\ncaller receives — where it is always ``false``, because CM-2 filters the\nothers out in the query. Emitting it unconditionally is what lets an admin\nUI mark a comment internal (CM-3) without a second response shape."
      },
...
      "Page_CommentOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CommentOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[CommentOut]"
      },
...
      "UserSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "id",
          "first_name",
          "last_name",
          "role"
        ],
        "title": "UserSummary",
        "description": "A user reference embedded in another resource (design.md §1)."
      },
      "ErrorDetail": {
        "title": "ErrorDetail",
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "title": "Code",
            "type": "string",
            "enum": [
              "BAD_REQUEST",
              "UNAUTHENTICATED",
              "FORBIDDEN",
              "NOT_FOUND",
              "CONFLICT",
              "VALIDATION_ERROR",
              "INTERNAL_ERROR"
            ],
            "description": "Machine-readable error class. Closed set."
          },
          "message": {
            "title": "Message",
            "type": "string",
            "description": "Human-readable summary, safe to show a user. Never carries server internals."
          },
          "fields": {
            "title": "Fields",
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "description": "Per-field messages, keyed by field name. Present only on `VALIDATION_ERROR`."
          }
        }
      },
      "ErrorEnvelope": {
        "title": "ErrorEnvelope",
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "description": "The one shape every error response in this API uses."
      }
    },
    "securitySchemes": {
      "accessTokenCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token",
        "description": "Session access token. Set by `POST /api/v1/auth/login` and `POST /api/v1/auth/refresh` as an `httpOnly`, `Secure`, `SameSite=Lax` cookie with a 1 hour lifetime. Sent automatically by the browser; not readable from JavaScript and not settable by a client."
      },
      "refreshTokenCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token",
        "description": "Opaque refresh token, 30 day lifetime, same cookie attributes as `access_token`. Read by `POST /api/v1/auth/refresh` only, which rotates it; replaying an already-rotated token revokes the whole family (AUTH-11)."
      }
    }
  }
}